LayoutTests/platform/mac-catalina/fast/text/international/system-language/navigator-language/navigator-language-fr-expected.txt

FAIL navigator.language should be fr-FR. Was fr.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
